chore: move applicationSettings and reportingPeriods into VueX modules; detangle them a little
