+++ src/views/Validation.vue
@@ -199,7 +199,7 @@ export default {
         }))
     },
     periodId: function () {
-      return this.$store.state.viewPeriodID
+      return this.$store.state.reportingPeriods.viewPeriodID
     }
   },
   methods: {
